podcast_rss_download_cloud_function/requirements.txt
functions-framework==3.*
requests
google-cloud-firestore
feedparser